+++ docs/changelog.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 Each revision is versioned by the date of the revision.
 
+## 2026-06-22
+
+- Improved test coverage for both `ingress` and `gateway-route` mode.
+
 ## 2026-06-11
 
 - Removed support for `gateway-route` v0 library.